requirements.txt
torch~=2.1.1
torchvision~=0.16.1
matplotlib~=3.8.2
pandas~=2.1.3
numpy~=1.26.2
Pillow~=10.1.0
scikit-learn
scikit-image
pyyaml
tqdm